src/facto/requirements.txt
python-telegram-bot>=21.0
openai>=1.0
anthropic>=0.18.0
python-dotenv>=1.0
pymongo>=4.0
aiohttp>=3.9.0
httpx>=0.25.0
